executer -> executor
it's the proper spelling
